Move code over to auxfun_videos.py
